+++ src/cli/index.ts
@@ -1,4 +1,4 @@
-export { CLIApplication } from './cli-application.js';
+export { CLIApplication } from './cli.application.js';
 export { CommandParser } from './command-parser.js';
 
 export { HelpCommand } from './commands/help.command.js';
