+++ src/rest/index.ts
@@ -1,1 +1,1 @@
-export { RestApplication } from './rest-application.js';
+export { RestApplication } from './rest.application.js';
