+++ src/shared/libs/file-reader/index.ts
@@ -1,2 +1,2 @@
 export { FileReader } from './file-reader.interface.js';
-export { TSVFileReader } from './tsv-file-reader.js';
+export { TSVFileReader } from './tsv.file-reader.js';
