+++ src/shared/libs/file-writer/index.ts
@@ -1,2 +1,2 @@
 export { FileWriter } from './file-writer.interface.js';
-export { TSVFileWriter } from './tsv-file-writer.js';
+export { TSVFileWriter } from './tsv.file-writer.js';
