+++ src/shared/libs/logger/index.ts
@@ -1,2 +1,2 @@
 export { Logger } from './logger.interface.js';
-export { PinoLogger } from './pino-logger.js';
+export { PinoLogger } from './pino.logger.js';
