+++ src/shared/libs/offer-generator/tsv.offer-generator.ts
